requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==37.6.0
idna==3.10
Markdown==3.9
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sentry-sdk==2.37.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
